bgpd: fix integer overflow in maximum prefix threshold check
bgp_maximum_prefix_overflow() computes the warning threshold as:

    pcount > (peer->pmax[afi][safi] * peer->pmax_threshold[afi][safi] / 100)

Both operands are unsigned integers (pmax is uint32_t, pmax_threshold
is uint8_t), so the multiplication is performed in uint32_t arithmetic.
When pmax exceeds ~16.7 million (2^24) and the threshold is high
(e.g. 75–100), the product overflows uint32_t, wrapping to a small
value. This makes the comparison almost always false, silently
disabling the prefix threshold warning.

RFC 7454 Section 8 recommends maximum prefix limits as a critical
defense mechanism against route table exhaustion and runaway BGP
sessions. An overflow that silently disables the threshold warning
undermines this protection: a peer can continue advertising prefixes
well past the configured warning threshold without any log message or
notification, delaying operator awareness until the hard limit is
reached (or the router runs out of memory).

Cast pmax to uint64_t before the multiplication so the product is
computed in 64-bit arithmetic, correctly handling prefix limits up to
the uint32_t maximum (~4 billion) with any threshold percentage.

+++ bgpd/bgp_route.c
@@ -5364,7 +5364,7 @@ bool bgp_maximum_prefix_overflow(struct peer *peer, afi_t afi, safi_t safi,
 			   PEER_STATUS_PREFIX_LIMIT);
 
 	if (pcount
-	    > (peer->pmax[afi][safi] * peer->pmax_threshold[afi][safi] / 100)) {
+	    > ((uint64_t)peer->pmax[afi][safi] * peer->pmax_threshold[afi][safi] / 100)) {
 		if (CHECK_FLAG(peer->af_sflags[afi][safi],
 			       PEER_STATUS_PREFIX_THRESHOLD)
 		    && !always)
